requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
typing-extensions>=4.9.0 